Move TopKElement to collections/
