+++ mlprogram/utils/__init__.py
@@ -1,3 +1,2 @@
 from mlprogram.utils.datatypes import Token, Query  # noqa
-from mlprogram.utils.top_k_element import TopKElement  # noqa
 from mlprogram.utils.functions import Compose, Sequence, Map, Flatten, Threshold, Pick, save, load  # noqa
